Prettier CSS for notifications in Drawer
